+++ package.json
@@ -3,29 +3,46 @@
     "version": "1.0.0",
     "description": "This server act as a middleware to communicate with Slack API.",
     "author": "rivernews",
-    "main": "index.js",
+    "main": "./dist/index.js",
     "scripts": {
-        "start": "node /usr/src/server/index.js",
-        "dev": "nodemon server/index.js",
+        "build": "tsc",
+        "dev": "ts-node server/index.ts",
+        "start": "nodemon ./dist/index.js",
+        "format": "prettier-standard '**/*.ts'",
         "proxy": "ngrok http 8080",
         "test": "mocha server/index.test.js --timeout 30000"
     },
     "dependencies": {
+        "@types/cheerio": "^0.22.16",
+        "@types/express": "^4.17.2",
         "axios": "^0.19.2",
         "cheerio": "^1.0.0-rc.3",
         "express": "^4.17.1",
-        "lodash": "^4.17.15"
+        "lodash": "^4.17.15",
+        "ts-node-dev": "^1.0.0-pre.44",
+        "typescript": "^3.7.5"
     },
     "devDependencies": {
         "chai": "^4.2.0",
         "husky": "^4.2.1",
+        "lint-staged": "^10.0.7",
         "mocha": "^7.0.1",
         "ngrok": "^3.2.7",
-        "nodemon": "^2.0.2"
+        "nodemon": "^2.0.2",
+        "prettier-standard": "^16.1.0"
     },
     "husky": {
         "hooks": {
+            "pre-commit": "lint-staged",
             "pre-push": "pwd && . ./devcontainer.env && export $(cut -d= -f1 ./devcontainer.env) || true && echo \"Port is running on ${PORT}\"  && npm test"
+        }
+    },
+    "lint-staged": {
+        "linters": {
+            "server/**/*.ts": [
+                "prettier-standard",
+                "git add"
+            ]
         }
     }
 }
