+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+        "allowJs": true,
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "pretty": true,
+        "sourceMap": true,
+        "target": "es5",
+        "strict": true,
+        "resolveJsonModule": true,
+        "esModuleInterop": true,
+        "outDir": "./dist",
+
+        "noImplicitAny": true
+    },
+    "include": [
+        "server/**/*.ts",
+        "server/**/*.js"
+    ],
+    "exclude": [
+        "node_modules"
+    ]
+}
